+++ src/mlpack/tests/random_forest_test.cpp
@@ -37,8 +37,8 @@ TEST_CASE("BootstrapNoWeightsTest", "[RandomForestTest]")
     arma::Row<size_t> bootstrapLabels;
     arma::rowvec bootstrapWeights;
 
-    Bootstrap<false>(dataset, labels, weights, bootstrapDataset,
-        bootstrapLabels, bootstrapWeights);
+    DefaultBootstrap().Bootstrap<false>(dataset, labels, weights,
+        bootstrapDataset, bootstrapLabels, bootstrapWeights);
 
     REQUIRE(bootstrapDataset.n_cols == 1000);
     REQUIRE(bootstrapDataset.n_rows == 1);
@@ -72,8 +72,8 @@ TEST_CASE("BootstrapWeightsTest", "[RandomForestTest]")
     arma::Row<size_t> bootstrapLabels;
     arma::rowvec bootstrapWeights;
 
-    Bootstrap<true>(dataset, labels, weights, bootstrapDataset,
-        bootstrapLabels, bootstrapWeights);
+    DefaultBootstrap().Bootstrap<true>(dataset, labels, weights,
+        bootstrapDataset, bootstrapLabels, bootstrapWeights);
 
     REQUIRE(bootstrapDataset.n_cols == 1000);
     REQUIRE(bootstrapDataset.n_rows == 1);
@@ -606,4 +606,141 @@ TEST_CASE("ExtraTreesAccuracyTest", "[RandomForestTest]")
   accuracy /= predictions.n_elem;
 
   REQUIRE(accuracy >= 0.85);
+}
+
+/**
+ * Test ComputeAverageUniqueness.
+ * 
+ * Average uniqueness is defined as average over the lifetime of an event
+ * of how many other events are active at every point in time.
+ */
+TEST_CASE("ComputeAverageUniquenessTest", "[RandomForestTest]")
+{
+  arma::umat indM(2 /* rows */, 3 /* cols */, arma::fill::zeros);
+
+  // indM = [1 1 1 0 0 0
+  //         0 0 1 1 0 0
+  //         0 0 0 0 1 1]
+  // The last event is fully isolated.
+  // The first two events overlap in the third column.
+  indM(0, 0) = 0;
+  indM(1, 0) = 2;
+  indM(0, 1) = 2;
+  indM(1, 1) = 3;
+  indM(0, 2) = 4;
+  indM(1, 2) = 5;
+
+  arma::vec concurrency(6);
+  concurrency[0] = 1;
+  concurrency[1] = 1;
+  concurrency[2] = 2;
+  concurrency[3] = 1;
+  concurrency[4] = 1;
+  concurrency[5] = 1;
+
+  const arma::vec avg(SequentialBootstrap<>::ComputeAverageUniqueness(indM,
+    arma::linspace<arma::uvec>(0, 2, 3), concurrency));
+
+  REQUIRE(avg.n_rows == 3);
+  REQUIRE(avg(0) == 5.0 / 6.0);
+  REQUIRE(avg(1) == 0.75);
+  REQUIRE(avg(2) == 1.0);
+}
+
+/**
+ * Test ComputeNextDrawProbabilities.
+ */
+TEST_CASE("ComputeNextDrawProbabilitiesTest", "[RandomForestTest]")
+{
+  arma::uvec phi1(2);
+  arma::umat indM(2 /* rows */, 3 /* cols */, arma::fill::zeros);
+
+  phi1[0] = 1u;
+
+  indM(0, 0) = 0;
+  indM(1, 0) = 2;
+  indM(0, 1) = 2;
+  indM(1, 1) = 3;
+  indM(0, 2) = 4;
+  indM(1, 2) = 5;
+
+  arma::vec concurrency(6);
+  concurrency[0] = 0;
+  concurrency[1] = 0;
+  concurrency[2] = 1;
+  concurrency[3] = 1;
+  concurrency[4] = 0;
+  concurrency[5] = 0;
+
+  // Compute the probabilities that observations 0, 1, 2 are drawn after
+  // observation 1 has already been drawn.
+  const arma::vec delta2(SequentialBootstrap<>::ComputeNextDrawProbabilities(
+    phi1, 1, concurrency, indM));
+
+  REQUIRE(delta2(0) == 5.0 / 14.0);
+  // Should have the lowest probability as it has already been drawn.
+  REQUIRE(delta2(1) == 3.0 / 14.0);
+  // Should have the highest probability as this event does not overlap
+  // with others.
+  REQUIRE(delta2(2) == 6.0 / 14.0);
+}
+
+/**
+ * Test ComputeSamples.
+ */
+TEST_CASE("ComputeSamplesTest", "[RandomForestTest]")
+{
+  arma::umat indM(2 /* rows */, 6 /* cols */, arma::fill::zeros);
+
+  indM(0, 0) = 0;
+  indM(1, 0) = 2;
+  indM(0, 1) = 2;
+  indM(1, 1) = 3;
+  indM(0, 2) = 4;
+  indM(1, 2) = 5;
+
+  indM(1, 3) = indM(1, 4) = indM(1, 5) = 5;
+
+  SequentialBootstrap bootstrap(indM);
+  const arma::uvec phi(bootstrap.ComputeSamples(6));
+
+  // Can only check that the next draw does not yield indices
+  // outside of the range of observations.
+  REQUIRE(phi(0) < 6);
+  REQUIRE(phi(1) < 6);
+  REQUIRE(phi(2) < 6);
+}
+
+/**
+ * Test SequentialBootstrap.
+ */
+TEST_CASE("SequentialBootstrapTest", "[RandomForestTest]")
+{
+  const arma::mat ds(10 /* rows */, 6 /* cols */, arma::fill::randu);
+  const arma::Row<size_t> labels{ 1, 0, 0, 0, 0, 0 };
+  const arma::vec weights(6, arma::fill::ones);
+  arma::umat indM(2 /* rows */, 6 /* cols */, arma::fill::zeros);
+
+  indM(0, 0) = 0;
+  indM(1, 0) = 2;
+  indM(0, 1) = 2;
+  indM(1, 1) = 3;
+  indM(0, 2) = 4;
+  indM(1, 2) = 5;
+
+  indM(1, 3) = indM(1, 4) = indM(1, 5) = 5;
+
+  SequentialBootstrap<> bootstrap(indM);
+  arma::mat bsDataset;
+  arma::Row<size_t> bsLabels;
+  arma::vec bsWeights;
+
+  bootstrap.Bootstrap<true>(ds, labels, weights, bsDataset,
+    bsLabels, bsWeights);
+
+  // Check that dimensions are the same.
+  REQUIRE(ds.n_rows == bsDataset.n_rows);
+  REQUIRE(ds.n_cols == bsDataset.n_cols);
+  REQUIRE(labels.n_cols == bsLabels.n_cols);
+  REQUIRE(weights.n_rows == bsWeights.n_rows);
 }