+++ HISTORY.md
@@ -10,6 +10,9 @@ _????-??-??_
 
  * Add `MLPACK_NO_STD_MUTEX` to allow disabling `std::mutex` (#3868).
 
+ * Added bootstrap strategies for RandomForest: IdentityBootstrap,
+   DefaultBootstrap, and SequentialBootstrap. (#3829)
+
 ## mlpack 4.5.1
 
 _2024-12-02_